dictate-app/sidecar/requirements.txt
numpy
scipy
sounddevice
moonshine-voice
torch
nemo_toolkit[asr]
